Migrate to Unified API, NUnitLite dependency update to c9e4c13f146ffa5bffd71bfdd91702ff6832b6a8, see https://github.com/spouliot/NUnitLite/commit/c9e4c13f146ffa5bffd71bfdd91702ff6832b6a8

+++ NUnitLite/TouchRunner/TouchOptions.cs
@@ -20,8 +20,8 @@
 
 using System;
 using MonoTouch.Dialog;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 using Mono.Options;
 
 namespace MonoTouch.NUnit.UI {
@@ -35,7 +35,7 @@ namespace MonoTouch.NUnit.UI {
 			var defaults = NSUserDefaults.StandardUserDefaults;
 			EnableNetwork = defaults.BoolForKey ("network.enabled");
 			HostName = defaults.StringForKey ("network.host.name");
-			HostPort = defaults.IntForKey ("network.host.port");
+			HostPort = (int)defaults.IntForKey ("network.host.port");
 			SortNames = defaults.BoolForKey ("display.sort");
 			
 			var os = new OptionSet () {
